Update readme and requirements-2.7.txt
The requirements-2.7.txt installs docx. When running the
raml2doc this results in an "ImportError: cannot import name
Document" error.

Using python-docx will prevent this error.

In the readme several minor typos were fixed.
Smart quotes were replaced with ASCII double quotes.
python2 was changed to python since that is the name
of the executable.

Signed-off-by: George Nash <george.nash@intel.com>